+++ release-notes.rst
@@ -2,6 +2,25 @@ Release notes for Chevah KeyCert
 ################################
 
 
+1.12.2 - 2020-07-09
+===================
+
+* Update command line help messages.
+
+
+1.12.1 - 2020-07-06
+===================
+
+* Allow defining key usage and constraints for Certificate Signing Requests
+  and self-signed certificates.
+
+
+1.12.0 - 2020-07-06
+===================
+
+* Allow creating self-signed certificates with custom attributes.
+
+
 1.11.1 - 2020-07-02
 ===================